image is compressed and retrieved

+++ Project/golom.h
@@ -1,6 +1,6 @@
 #include "mybitstream.h"
 
 
-void golom_encode(uchar codeword,short m,Bitblock &bitblock);
-uchar golom_decode(short m,Bitblock &bitblock); 
+void golom_encode(int codeword,short m,Bitblock &bitblock);
+int golom_decode(short m,Bitblock &bitblock);
